+++ Functions/Assertions/PesterThrow.ps1
@@ -1,52 +1,122 @@
-function PesterThrow([scriptblock] $ActualValue, $ExpectedMessage, $ErrorId, [switch] $Negate) {
-    $script:ActualExceptionMessage = ""
-    $script:ActualExceptionWasThrown = $false
+function PesterThrow([scriptblock] $ActualValue, $ExpectedMessage, $ErrorId, [type]$ExceptionType, [switch] $Negate, [string] $Because, [switch] $PassThru) {
+    $actualExceptionMessage = ""
+    $actualExceptionWasThrown = $false
+    $actualError = $null
+    $actualException = $null
+    $actualExceptionLine = $null
 
     if ($null -eq $ActualValue) {
         throw (New-Object -TypeName ArgumentNullException -ArgumentList "ActualValue","Scriptblock not found. Input to 'Throw' and 'Not Throw' must be enclosed in curly braces.")
     }
-
-    # This is superfluous, here for now.
-    $ExpectedErrorId = $ErrorId
 
     try {
         do {
             $null = & $ActualValue
         } until ($true)
     } catch {
-        $script:ActualExceptionWasThrown = $true
-        $script:ActualExceptionMessage = $_.Exception.Message
-        $script:ActualErrorId = $_.FullyQualifiedErrorId
-        $script:ActualExceptionLine = Get-ExceptionLineInfo $_.InvocationInfo
+        $actualExceptionWasThrown = $true
+        $actualError = $_
+        $actualException = $_.Exception
+        $actualExceptionMessage = $_.Exception.Message
+        $actualErrorId = $_.FullyQualifiedErrorId
+        $actualExceptionLine = (Get-ExceptionLineInfo $_.InvocationInfo) -replace [System.Environment]::NewLine,"$([System.Environment]::NewLine)    "
     }
 
     [bool] $succeeded = $false
 
-    if ($ActualExceptionWasThrown) {
-        $succeeded = (Get-DoValuesMatch $script:ActualExceptionMessage $ExpectedMessage) -and
-                     (Get-DoValuesMatch $script:ActualErrorId $ExpectedErrorId)
+    if ($Negate) {
+        # this is for Should -Not -Throw. Once *any* exception was thrown we should fail the assertion
+        # there is no point in filtering the exception, because there should be none
+        $succeeded = -not $actualExceptionWasThrown
+        if (-not $succeeded) {
+            $failureMessage = "Expected no exception to be thrown,$(Format-Because $Because) but an exception was thrown $actualExceptionLine."
+            return New-Object psobject -Property @{
+                Succeeded      = $succeeded
+                FailureMessage = $failureMessage
+            }
+        } else {
+            return New-Object psobject -Property @{
+                Succeeded      = $true
+            }
+        }
     }
 
-    if ($Negate) { $succeeded = -not $succeeded }
+    # the rest is for Should -Throw, we must fail the assertion when no exception is thrown
+    # or when the exception does not match our filter
 
-    $failureMessage = ''
+    function Join-And ($Items, $Threshold=2) {
 
-    if (-not $succeeded)
-    {
-        if ($Negate)
+        if ($null -eq $items -or $items.count -lt $Threshold)
         {
-            $failureMessage = NotPesterThrowFailureMessage -ActualValue $ActualValue -ExpectedMessage $ExpectedMessage -ExpectedErrorId $ExpectedErrorId
+            $items -join ', '
         }
         else
         {
-            $failureMessage = PesterThrowFailureMessage -ActualValue $ActualValue -ExpectedMessage $ExpectedMessage -ExpectedErrorId $ExpectedErrorId
+            $c = $items.count
+            ($items[0..($c-2)] -join ', ') + ' and ' + $items[-1]
         }
     }
 
-    return New-Object psobject -Property @{
-        Succeeded      = $succeeded
-        FailureMessage = $failureMessage
+    function Add-SpaceToNonEmptyString ([string]$Value) {
+        if ($Value)
+        {
+            " $Value"
+        }
     }
+
+    $buts = @()
+    $filters = @()
+
+    $filterOnExceptionType = $null -ne $ExceptionType
+    if ($filterOnExceptionType) {
+        $filters += "with type {$ExceptionType}"
+
+        if ($actualExceptionWasThrown -and $actualException -isnot $ExceptionType) {
+            $buts += "the exception type was '{$($actualException.GetType())}'"
+        }
+    }
+
+    $filterOnMessage = -not [string]::IsNullOrEmpty($ExpectedMessage -replace "\s")
+    if ($filterOnMessage) {
+        $filters += "with message '$ExpectedMessage'"
+        if ($actualExceptionWasThrown -and (-not (Get-DoValuesMatch $actualExceptionMessage $ExpectedMessage))) {
+            $buts += "the message was '$($actualExceptionMessage)'"
+        }
+    }
+
+    $filterOnId = -not [string]::IsNullOrEmpty($ErrorId -replace "\s")
+    if ($filterOnId) {
+        $filters += "with FullyQualifiedErrorId '$ErrorId'"
+        if ($actualExceptionWasThrown -and (-not (Get-DoValuesMatch $actualErrorId $ErrorId))) {
+            $buts += "the FullyQualifiedErrorId was '$($actualErrorId)'"
+        }
+    }
+
+    if (-not $actualExceptionWasThrown)
+    {
+        $buts += "no exception was thrown"
+    }
+
+    if ($buts.Count -ne 0) {
+        $filter = Add-SpaceToNonEmptyString ( Join-And $filters -Threshold 3 )
+        $but = Join-And $buts
+        $failureMessage = "Expected an exception,$filter to be thrown,$(Format-Because $Because) but $but. $actualExceptionLine".Trim()
+
+        return New-Object psobject -Property @{
+            Succeeded      = $false
+            FailureMessage = $failureMessage
+        }
+    }
+
+    $result = New-Object psobject -Property @{
+        Succeeded      = $true
+    }
+
+    if ($PassThru) {
+        $result | Add-Member -MemberType NoteProperty -Name 'Data' -Value $actualError
+    }
+
+    return $result
 }
 
 function Get-DoValuesMatch($ActualValue, $ExpectedValue) {
@@ -58,68 +128,16 @@ function Get-DoValuesMatch($ActualValue, $ExpectedValue) {
 
 function Get-ExceptionLineInfo($info) {
     # $info.PositionMessage has a leading blank line that we need to account for in PowerShell 2.0
-    $positionMessage = $info.PositionMessage -split '\r?\n' -match '\S' -join "`r`n"
+    $positionMessage = $info.PositionMessage -split '\r?\n' -match '\S' -join [System.Environment]::NewLine
     return ($positionMessage -replace "^At ","from ")
 }
 
-function PesterThrowFailureMessage($ActualValue, $ExpectedMessage, $ExpectedErrorId) {
-    $StringBuilder = Microsoft.PowerShell.Utility\New-Object System.Text.StringBuilder
-    $null = $StringBuilder.Append('Expected: the expression to throw an exception')
-
-    if ($ExpectedMessage -or $ExpectedErrorId)
-    {
-        $null = $StringBuilder.Append(' with ')
-        $Expected = switch ($null)
-        {
-            { $ExpectedMessage } { 'message {{{0}}}' -f $ExpectedMessage }
-            { $ExpectedErrorId } { 'error id {{{0}}}' -f $ExpectedErrorId }
-        }
-        $Actual = switch ($null)
-        {
-            { $ExpectedMessage } { 'message was {{{0}}}' -f $ActualExceptionMessage }
-            { $ExpectedErrorId } { 'error id was {{{0}}}' -f $ActualErrorId }
-        }
-        $null = $StringBuilder.Append(("{0}, an exception was {1}raised, {2}$([System.Environment]::NewLine)    {3}" -f
-            ($Expected -join ' and '),
-            @{$true="";$false="not "}[$ActualExceptionWasThrown],
-            ($Actual -join ' and '),
-            ($ActualExceptionLine -replace "`n","$([System.Environment]::NewLine)    ")
-        ))
-    }
-
-    return $StringBuilder.ToString()
+function PesterThrowFailureMessage {
+    # to make the should tests happy, for now
 }
 
-function NotPesterThrowFailureMessage($ActualValue, $ExpectedMessage, $ExpectedErrorId) {
-    $StringBuilder = New-Object System.Text.StringBuilder
-    $null = $StringBuilder.Append('Expected: the expression not to throw an exception')
-
-    if ($ExpectedMessage -or $ExpectedErrorId)
-    {
-        $null = $StringBuilder.Append(' with ')
-        $Expected = switch ($null)
-        {
-            { $ExpectedMessage } { 'message {{{0}}}' -f $ExpectedMessage }
-            { $ExpectedErrorId } { 'error id {{{0}}}' -f $ExpectedErrorId }
-        }
-        $Actual = switch ($null)
-        {
-            { $ExpectedMessage } { 'message was {{{0}}}' -f $ActualExceptionMessage }
-            { $ExpectedErrorId } { 'error id was {{{0}}}' -f $ActualErrorId }
-        }
-        $null = $StringBuilder.Append(("{0}, an exception was {1}raised, {2}$([System.Environment]::NewLine)    {3}" -f
-            ($Expected -join ' and '),
-            (@{$true="";$false="not "}[$ActualExceptionWasThrown]),
-            ($Actual -join ' and '),
-            ($ActualExceptionLine  -replace "$([System.Environment]::NewLine)","$([System.Environment]::NewLine)    ")
-        ))
-    }
-    else
-    {
-      $null = $StringBuilder.Append((". Message was {{{0}}}$([System.Environment]::NewLine)    {1}" -f $ActualExceptionMessage, ($ActualExceptionLine -replace "$([System.Environment]::NewLine)","$([System.Environment]::NewLine)    ")))
-    }
-
-    return $StringBuilder.ToString()
+function NotPesterThrowFailureMessage {
+    # to make the should tests happy, for now
 }
 
 Add-AssertionOperator -Name Throw `