+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 4.2.0-alpha3 (December 17, 2017)
+- Add -Because parameters to all assertions [GH-959]
+- Add -BeLessOrEqual and -BeGreaterOrEqual
+- Add -Contain (that operates on arrays)
+- Add -BeLikeExactly
+- Add -HaveType alias to -BeOfType
+- Fix assertion messages in -BeOfType
+- Throw argument exception when -BeOfType is given type that is not loaded
+- Add -PassThru to -Throw to get the exception when some is thrown and passes the filters
+- Add -BeTrue to test for truthy values
+- Add -BeFalse to test for falsy values
+- Add -HaveCount to count stuff in collections
+
+## 4.2.0-alpha2 (December 12, 2017)
+   - Load dependencies optionally, because they are not part of the package build
+
+## 4.2.0-alpha1 (December 12, 2017)
+   - Should Throw filters on exception type [GH-954]
+
+## 4.1.1 (December 09, 2017)
+  - Fix deployment scripts and package on choco and nuget
+  
 ## 4.1.0 (November 15, 2017)
   - Help for the Assert-VerifiableMocks function added [GH-883]
   - Add-AssertionOperator can be called multiple times for identical parameters without errors.  [GH-893]